Added @ConfigurationPropertyValue annotation and its processing (#7986).
Added onNull() to BindHandler to be able to react to properties that
bind to null to choose a fallback value instead.

ConfigurationPropertyValue annotations (which allow defining a fallback
value for a property) are now collected by
ConfigurationPropertyValueReader and a map of fallback properties is
passed into a FallbackBindHandler. The handler then reacts on a null
value by providing the fallback value instead of the null value.